feat: support floating-point number literals in filter expressions
- Change TokenKind::Number from i64 to f64 in lexer
- Update read_number() to handle decimal points and exponent notation
- Change Literal::Number from i64 to f64 in AST
- Update parser and evaluator to handle f64 numbers
- Add lexer tests for float literals (decimal, exponent, negative)
- Add eval tests for float comparisons in filter expressions

+++ crates/jpp_core/src/ast.rs
@@ -101,8 +101,8 @@ pub enum Literal {
     Null,
     /// Boolean value
     Bool(bool),
-    /// Integer number
-    Number(i64),
+    /// Number (integer or floating-point)
+    Number(f64),
     /// String value
     String(String),
 }
